Review response.
Comments.

Move ConstPropagationAndUnrolling.

Rename to StaticAnalyzing.
