+++ compiler/compiler/src/compiler.rs
@@ -54,47 +54,6 @@ pub struct Compiler<N: Network> {
     phantom: std::marker::PhantomData<N>,
 }
 
-/// Pass that runs const propagation and loop unrolling until a fixed point.
-struct ConstPropagationAndUnrolling;
-
-impl Pass for ConstPropagationAndUnrolling {
-    type Input = ();
-    type Output = ();
-
-    const NAME: &str = "ConstPropagationAndUnrolling";
-
-    fn do_pass(_input: Self::Input, state: &mut CompilerState) -> Result<Self::Output> {
-        const LARGE_LOOP_BOUND: usize = 1024usize;
-
-        for _ in 0..LARGE_LOOP_BOUND {
-            let loop_unroll_output = Unrolling::do_pass((), state)?;
-
-            let const_prop_output = ConstPropagation::do_pass((), state)?;
-
-            if !const_prop_output.changed && !loop_unroll_output.loop_unrolled {
-                // We've got a fixed point, so see if we have any errors.
-                if let Some(not_evaluated_span) = const_prop_output.const_not_evaluated {
-                    return Err(CompilerError::const_not_evaluated(not_evaluated_span).into());
-                }
-
-                if let Some(not_evaluated_span) = const_prop_output.array_index_not_evaluated {
-                    return Err(CompilerError::array_index_not_evaluated(not_evaluated_span).into());
-                }
-
-                if let Some(not_unrolled_span) = loop_unroll_output.loop_not_unrolled {
-                    return Err(CompilerError::loop_bounds_not_evaluated(not_unrolled_span).into());
-                }
-
-                return Ok(());
-            }
-        }
-
-        // Note that it's challenging to write code in practice that demonstrates this error, because Leo code
-        // with many nested loops or operations will blow the stack in the compiler before this bound is hit.
-        Err(CompilerError::const_prop_unroll_many_loops(LARGE_LOOP_BOUND, Default::default()).into())
-    }
-}
-
 impl<N: Network> Compiler<N> {
     pub fn parse(&mut self, source: &str, filename: FileName) -> Result<()> {
         // Register the source in the source map.
@@ -160,8 +119,8 @@ impl<N: Network> Compiler<N> {
             &mut self.state,
         )?;
 
-        StaticAnalysis::do_pass(
-            StaticAnalysisInput {
+        StaticAnalyzing::do_pass(
+            StaticAnalyzingInput {
                 max_depth: self.compiler_options.build.conditional_block_max_depth,
                 conditional_branch_type_checking: !self.compiler_options.build.disable_conditional_branch_type_checking,
             },
